requirements.txt
opencv-python>=4.5.5
face-recognition>=1.3.0
dlib>=19.24.0
numpy>=1.22.3
streamlit>=1.8.1
Pillow>=9.1.0
matplotlib>=3.5.1
